Removed excess code
